move InitContext to separate file; add [Must]Wait* methods.
